[Publishing] stops onScroll event from re-rendering article
